requirements.txt
jsonrpclib
zbar
opencv-python
Pillow